tests/submission/Comcast/title_element.html

<!DOCTYPE HTML>
<title>Title	 Test</title>

<h1>Title element test</h1>

<div id="log"></div>
<script src="/resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<script>

test(function() {assert_equals(document.title, "Title Test")}, "The IDL attribute text must return a concatenation of the contents of all the text nodes that are children of the title element (ignoring any other nodes such as comments or elements), in tree order. ");

</script>
